Integrated Update and Create
Integrated the update and create functionality to support the use call
to the controller from a single mms queue that has no understanding of
workflow. To support this we make an "exists" query on the index first.
Secondly, took out the crazy Bas64 index id. Finally prepended the
record id with the index id.